WebApplication Example
I got an template from the web, so don´t you think I did all the web
design and javascript stuff. I, however, had to make it to work in the
ASP.NET environment, as the template was only one big HTML file along
with CSS and JS. I created a Master Web Page (to keep it simple, its
basically like a Master Slide from Power Point), so in the future it
will be easier to make big changes in design or whatsoever.

I didn´t have too much time to fix some Javascript stuffs (animations)
and till now there is only a simple functionality that simulates a
Search Request.