Context Selectors Added
Just an update. Now it is possible to select one or more contexts from
the buttons on the right side. Following Thomas idea, I tried to make it
like the Tiles of Windows 8. Of course a lot more needs to be done to
make it look like that, but for now I think its good.